requirements.txt
dgl==1.0.0
dgl_cu102==0.6.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
scikit_learn==1.2.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.64.1